+++ korus-ui/components/FileList/theme.ts
@@ -1,0 +1,10 @@
+/* eslint-disable key-spacing */
+export const defaultFileListTheme = {
+  fileList:               'file-list table',
+  fileListDeleteIcon:     'file-list-delete',
+  fileListDownloadIcon:   'file-list-download',
+  fileListTag:            'file-list-tag',
+  fileListWrapper:        'file-list-wrapper',
+  subtitle:               'file-list-subtitle',
+};
+/* eslint-enable key-spacing */